fix: auth - delete unused files
