+++ src/auth/auth.module.ts
@@ -2,8 +2,8 @@ import { UserModule } from './../users/users.module';
 import { JwtStrategy } from './strategies/jwt.strategy.';
 
 import { Module } from '@nestjs/common';
-import { JwtModule } from '@nestjs/jwt/dist';
-import { PassportModule } from '@nestjs/passport/dist';
+import { JwtModule } from '@nestjs/jwt';
+import { PassportModule } from '@nestjs/passport';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { LocalStrategy } from './strategies/local.strategy';
