+++ src/auth/guards/jwt-auth.guard.ts
@@ -1,4 +1,8 @@
-import { ExecutionContext, Injectable } from '@nestjs/common';
+import {
+  ExecutionContext,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import { AuthGuard } from '@nestjs/passport';
 import { IS_PUBLIC_KEY } from '../decorators/is-public.decorator';
@@ -9,13 +13,25 @@ export class JwtAuthGuard extends AuthGuard('jwt') {
     super();
   }
 
-  canActivate(context: ExecutionContext): Promise<boolean> | boolean {
+  async canActivate(context: ExecutionContext): Promise<boolean> {
     const isPublic = this.reflector.getAllAndOverride<boolean>(IS_PUBLIC_KEY, [
+      context.getHandler(),
       context.getClass(),
     ]);
 
     if (isPublic) {
       return true;
     }
+
+    const canActivate = super.canActivate(context);
+    if (typeof canActivate === 'boolean') {
+      return canActivate;
+    }
+
+    const canActivatePromise = canActivate as Promise<boolean>;
+
+    return canActivatePromise.catch(() => {
+      throw new UnauthorizedException();
+    });
   }
 }
