+++ src/events/dto/create-event-images.dto.ts
@@ -9,5 +9,5 @@ export class CreateEventsImages {
     },
     description: 'This field accepts images with JPG/JPEG/PNG and GIFT format',
   })
-  files: any[];
+  files: Express.Multer.File[];
 }
